backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyMuPDF
sentence-transformers
faiss-cpu
